+++ package.json
@@ -18,10 +18,16 @@
     "react-dom": "18.2.0",
     "react-native": "0.71.6",
     "react-native-dotenv": "^3.4.8",
-    "react-native-web": "~0.18.10"
+    "react-native-web": "~0.18.10",
+    "typescript": "^4.9.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0"
+    "@babel/core": "^7.20.0",
+    "@tsconfig/react-native": "^3.0.0",
+    "@types/jest": "^29.5.1",
+    "@types/react": "^18.0.38",
+    "@types/react-test-renderer": "^18.0.0",
+    "typescript": "^4.9.4"
   },
   "private": true
 }
