+++ screens/HomeScreen.tsx
@@ -1,0 +1,59 @@
+import React from "react";
+import { StyleSheet, Text, View, TouchableOpacity } from "react-native";
+import { auth } from "../firebase";
+import { useNavigation } from "@react-navigation/core";
+import { HomeProps } from "../types";
+
+const HomeScreen = (props: HomeProps) => {
+  const navigation = useNavigation<(typeof props)["navigation"]>();
+
+  const handleLogout = () => {
+    auth
+      .signOut()
+      .then(() => {
+        navigation.replace("Login");
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Home Screen</Text>
+      <Text>Email: {auth.currentUser?.email}</Text>
+      <TouchableOpacity onPress={handleLogout} style={styles.button}>
+        <Text style={styles.buttonText}>Sign out</Text>
+      </TouchableOpacity>
+    </View>
+  );
+};
+
+export default HomeScreen;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "blue",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  title: {
+    fontSize: 50,
+    fontWeight: "bold",
+  },
+  button: {
+    backgroundColor: "blue",
+    color: "white",
+    padding: 10,
+    margin: 10,
+    borderRadius: 5,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  buttonText: {
+    color: "white",
+    fontSize: 20,
+    fontWeight: "bold",
+  },
+});
