--- screens/LoginScreen.js
@@ -1,131 +1,0 @@
-import React, { useEffect, useState } from "react";
-import { StyleSheet, TextInput, Text, View, KeyboardAvoidingView, TouchableOpacity } from "react-native";
-import { auth } from "../firebase";
-import { createUserWithEmailAndPassword, signInWithEmailAndPassword } from "firebase/auth";
-import { useNavigation } from "@react-navigation/core";
-
-const LoginScreen = () => {
-    const [email, setEmail] = useState("");
-    const [password, setPassword] = useState("");
-
-    const navigation = useNavigation();
-
-    useEffect(() => {
-        const unsibscribe = auth.onAuthStateChanged((user) => {
-            if (user) {
-                navigation.replace("Home");
-            }
-        });
-        return unsibscribe;
-    }, []);
-
-    const handleSignUp = async () => {
-        try {
-            const userCredentials = await createUserWithEmailAndPassword(auth, email, password);
-            const user = userCredentials.user;
-            console.log("Registered with:", user.email);
-        }
-        catch (error) {
-            const errorCode = error.code;
-            const errorMessage = error.message;
-            console.log("Error Code: ", errorCode);
-            console.log("Error Message: ", errorMessage);
-        }     
-    }
-
-    const handleLogin = async () => {
-        try {
-            const userCredentials = await signInWithEmailAndPassword(auth, email, password);
-            const user = userCredentials.user;
-            console.log("Logged in with:", user.email);
-        }
-        catch (error) {
-            const errorCode = error.code;
-            const errorMessage = error.message;
-            console.log("Error Code: ", errorCode);
-            console.log("Error Message: ", errorMessage);
-        }
-    }
-
-    return (
-        <KeyboardAvoidingView 
-            style={styles.container} 
-            behavior="padding"
-            >
-            <Text style={styles.title}>Login Screen</Text>
-            <View style={styles.inputContainer}>
-                <TextInput 
-                    placeholder="Email" 
-                    value={ email } 
-                    onChangeText={text => setEmail(text)} 
-                    style={styles.input}
-                />
-                <TextInput 
-                    placeholder="Password" 
-                    value={ password } 
-                    onChangeText={text => setPassword(text)} 
-                    style={styles.input}
-                    secureTextEntry
-                />
-            </View>  
-            <View>
-                <TouchableOpacity
-                    onPress={handleLogin}
-                    style={styles.button}
-                >
-                    <Text style={styles.buttonText}>Login</Text>
-                </TouchableOpacity>
-                <TouchableOpacity
-                    onPress={handleSignUp}
-                    style={styles.button}
-                >
-                    <Text style={styles.buttonText}>Register</Text>
-                </TouchableOpacity>
-            </View>
-        </KeyboardAvoidingView>
-    );
-}
-
-export default LoginScreen;
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: "center",
-        justifyContent: "center",
-    },
-    title: {
-        fontSize: 40,
-        fontWeight: "bold",
-        marginBottom: 20,
-    },
-    inputContainer: {
-        width: 300,
-        borderWidth: 1,
-        borderStyle: "solid",
-        borderColor: "black",
-        padding: 10,
-    },
-    input: {
-        height: 40,
-        margin: 12,
-        borderWidth: 1,
-        borderStyle: "solid",
-        borderColor: "black",
-        padding: 10,
-    },
-    button: {
-        backgroundColor: "blue",
-        color: "white",
-        padding: 10,
-        margin: 10,
-        borderRadius: 5,
-        alignItems: "center",
-        justifyContent: "center",
-    },
-    buttonText: {
-        color: "white",
-        fontSize: 20,
-        fontWeight: "bold",
-    }
-});
