+++ types.tsx
@@ -1,0 +1,9 @@
+import type { NativeStackScreenProps } from "@react-navigation/native-stack";
+
+export type RootStackParamList = {
+  Home: undefined;
+  Login: undefined;
+};
+
+export type LoginProps = NativeStackScreenProps<RootStackParamList, "Login">;
+export type HomeProps = NativeStackScreenProps<RootStackParamList, "Home">;
